rename: compare-internet-packages -> besarfeh (به‌صرفه)
Catchier name meaning "cost-effective / good value" in Persian — exactly
the tool's promise. Renames the package (src/besarfeh/), the console
command (`besarfeh`), the distribution name, and all references in
docs/README/CLAUDE.md/tests. uv.lock regenerated. Behavior unchanged.

NOTE: the GitHub repo itself still needs renaming (Settings -> rename, or
`gh repo rename besarfeh`) + remote URL update — left to the maintainer.
