+++ src/besarfeh/scrapers/rightel.py
